requirements.txt
streamlit==1.41.1
pandas==2.2.3
numpy==2.1.3
scikit-learn==1.5.2
requests==2.32.3
Pillow==11.0.0
torch==2.5.1
torchvision==0.20.1
